+++ src/app.js
@@ -2,10 +2,6 @@ const express = require('express');
 
 const app = express();
 
-
-
-
-
 // INCLUINDO ROTA
 
 const detranRouter = require('./routes/detran');
@@ -16,13 +12,11 @@ const indexRouter = require('./routes/index');
 
 // INCLUINDO MIDDLEWARE
 
-app.use(express.json());
+app.use(express.json(),
+    detranRouter,
+    pmRouter,
+    indexRouter);
 
-app.use('/detran', detranRouter);
-
-app.use('/pm', pmRouter);
-
-app.use('/', indexRouter);
