+++ src/controllers/indexController.js
@@ -1,0 +1,5 @@
+mensagemAPI = (req, res, next) => {
+    res.json("API de busca de carros");
+}
+
+module.exports = { mensagemAPI }
